requirements.txt
asgiref==3.2.7
Django==2.2
django-debug-toolbar==2.2
django-filter==2.2.0
django-haystack==2.8.1
django-pure-pagination==0.3.0
django-redis==4.11.0
djangorestframework==3.11.0
elasticsearch==2.4.1
Faker==4.0.3
Markdown==3.2.1
mysqlclient==1.4.6
python-dateutil==2.8.1
pytz==2019.3
redis==3.4.1
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.3
urllib3==1.25.9
